Update objReaderCSharpV4.cs
I am loading relatively small meshes and noticed that this change is upwards of 2x faster, with least noticeable difference being the simplest meshes (EX: a cube). I don't have anything to test larger meshes with, but figured I'd pass this along in case you were interested in making the changes. Fair warning, I made significant changes to how mine works, so there is a good chance that this doesnt compile, but the idea is still there.

Meshes and UVs and the like are way out of my comfort range, don't suppose you could enlighten me on why the Z axis isn't included in your calculations? I attempted to add it and noticed a significantly longer load time (3x or more), is this the only reason? As stated, adding Z axis with my version is almost unnoticeable.